refactor(vault): connections_service を導入 (#782 R4)
commands/vault.rs (650 行) の 10 コマンドに反復していた
「load → find(id) → mutate → save」を update_connection に集約し、
コマンド層を「main ウィンドウ検証 + service 呼び出し」のみに縮退。

- exposed↔trust 連動 (開示解除でクラス trust 剥奪 + Plugin は個体 trust
  全消し #712 §6.1)・trust の冪等性・プラグイン名スナップショット更新を
  Connection 上の純関数 (apply_*) に切り出し、直接ユニットテストで固定
- upsert/secret/削除/疎通テスト/AI プロバイダー移行の手続きも service へ。
  TOCTOU 回避や「secret 先消し」の順序根拠はコメントごと維持
- validate_base_url にもテスト追加。コマンド doc 短縮に伴い bindings 再生成
