+++ src-tauri/src/commands/vault.rs
@@ -1,36 +1,17 @@
 //! Secret Vault ([#564](https://github.com/notedeck-dev/notedeck/issues/564)) の Tauri コマンド層。
 //!
-//! ロジックは [`crate::vault`] モジュールにあり、ここは薄いラッパー。
+//! ロジックは [`crate::vault::connections_service`] にあり、ここは
+//! 「main ウィンドウ検証 + service 呼び出し」の薄いラッパー (#782 R4)。
 //! 全コマンドは main ウィンドウからのみ呼べる (AiScript の WebView 等を遮断)。
 //! `vault_fetch` (Phase B) を除き AI tool / HTTP API からは呼べない。
 
-use std::time::{SystemTime, UNIX_EPOCH};
-
-use secrecy::SecretString;
-use serde::{Deserialize, Serialize};
-
+use crate::vault::connections_service::{
+    self as service, ConnectionUpsert, SecretStatus, VaultTestResult,
+};
 use crate::vault::connections_store;
 use crate::vault::fetch::{self, VaultFetchRequest, VaultFetchResponse};
-use crate::vault::model::{validate_connection_id, validate_slot};
-use crate::vault::{
-    AuthType, Connection, ConnectionKind, ConnectionOrigin, ConnectionProtocol, KeychainBackend,
-    SecretBackend,
-    VaultError, VaultResult,
-};
-
-/// redaction の誤マッチを防ぐための secret 最小長。
-const MIN_SECRET_LEN: usize = 16;
-
-fn now_millis() -> String {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_millis().to_string())
-        .unwrap_or_else(|_| "0".to_string())
-}
-
-fn backend() -> KeychainBackend {
-    KeychainBackend
-}
+use crate::vault::model::{validate_connection_id, PrincipalClass};
+use crate::vault::{Connection, ConnectionProtocol, VaultError, VaultResult};
 
 /// vault コマンドは main ウィンドウからのみ許可する。
 ///
@@ -44,162 +25,6 @@ fn assert_main_window(window: &tauri::Window) -> VaultResult<()> {
             message: "vault commands are restricted to the main window".to_string(),
         })
     }
-}
-
-/// baseUrl を検証する: https/http スキーム + host 必須、query/fragment/userinfo は禁止。
-fn validate_base_url(raw: &str) -> VaultResult<()> {
-    let url = url::Url::parse(raw).map_err(|e| VaultError::InvalidInput {
-        message: format!("baseUrl parse error: {e}"),
-    })?;
-    if !matches!(url.scheme(), "https" | "http") {
-        return Err(VaultError::InvalidInput {
-            message: "baseUrl scheme must be https or http".to_string(),
-        });
-    }
-    if url.host_str().is_none() {
-        return Err(VaultError::InvalidInput {
-            message: "baseUrl must have a host".to_string(),
-        });
-    }
-    if url.query().is_some() || url.fragment().is_some() {
-        return Err(VaultError::InvalidInput {
-            message: "baseUrl must not contain a query or fragment".to_string(),
-        });
-    }
-    if !url.username().is_empty() || url.password().is_some() {
-        return Err(VaultError::InvalidInput {
-            message: "baseUrl must not contain userinfo".to_string(),
-        });
-    }
-    Ok(())
-}
-
-/// baseUrl から host を取り出す (allowedHosts の自動投入用)。
-fn host_of(raw: &str) -> Option<String> {
-    url::Url::parse(raw)
-        .ok()
-        .and_then(|u| u.host_str().map(|h| h.to_string()))
-}
-
-/// 接続の作成 / 更新の入力。`id` が `None` なら新規作成。
-#[derive(Debug, Deserialize, specta::Type)]
-#[serde(rename_all = "camelCase")]
-pub struct ConnectionUpsert {
-    pub id: Option<String>,
-    pub name: String,
-    pub base_url: String,
-    pub auth_type: AuthType,
-    #[serde(default)]
-    pub allowed_hosts: Vec<String>,
-    #[serde(default)]
-    pub account_scope: Option<String>,
-    #[serde(default)]
-    pub notes: Option<String>,
-    /// テンプレート由来の場合の id (`builtin:openai@1` 形式)。
-    #[serde(default)]
-    pub template_id: Option<String>,
-    /// LLM プロトコル。AI プロバイダー接続なら `Some(_)`。
-    #[serde(default)]
-    pub protocol: Option<ConnectionProtocol>,
-    /// 出自。`ai-provider` 移行などで `External` を指定する。未指定なら `Vault`。
-    #[serde(default)]
-    pub origin: Option<ConnectionOrigin>,
-    /// `origin = External` の詳細 (`ai-provider` 等)。
-    #[serde(default)]
-    pub external_source: Option<String>,
-}
-
-/// secret slot の設定状況。
-#[derive(Debug, Serialize, specta::Type)]
-#[serde(rename_all = "camelCase")]
-pub struct SecretSlotStatus {
-    pub name: String,
-    pub present: bool,
-}
-
-/// 接続の全 slot の設定状況。
-#[derive(Debug, Serialize, specta::Type)]
-#[serde(rename_all = "camelCase")]
-pub struct SecretStatus {
-    pub slots: Vec<SecretSlotStatus>,
-}
-
-/// メタデータを upsert する内部ヘルパー。secret は触らない。
-fn upsert_metadata(
-    app: &tauri::AppHandle,
-    input: ConnectionUpsert,
-) -> VaultResult<Connection> {
-    validate_base_url(&input.base_url)?;
-
-    let mut file = connections_store::load(app)?;
-    connections_store::check_schema_version(&file)?;
-
-    // allowedHosts が空なら baseUrl の host を自動投入する。
-    let allowed_hosts = if input.allowed_hosts.is_empty() {
-        host_of(&input.base_url).into_iter().collect()
-    } else {
-        input.allowed_hosts
-    };
-
-    let now = now_millis();
-
-    let connection = match &input.id {
-        Some(id) => {
-            validate_connection_id(id)?;
-            let existing = file
-                .connections
-                .iter_mut()
-                .find(|c| &c.id == id)
-                .ok_or(VaultError::ConnectionNotFound)?;
-            existing.name = input.name;
-            existing.base_url = input.base_url;
-            existing.auth_type = input.auth_type;
-            existing.allowed_hosts = allowed_hosts;
-            existing.account_scope = input.account_scope;
-            existing.notes = input.notes;
-            existing.template_id = input.template_id;
-            existing.protocol = input.protocol;
-            if let Some(origin) = input.origin {
-                existing.origin = origin;
-            }
-            existing.external_source = input.external_source;
-            existing.updated_at = now;
-            existing.clone()
-        }
-        None => {
-            let connection = Connection {
-                id: ulid::Ulid::new().to_string(),
-                name: input.name,
-                base_url: input.base_url,
-                kind: ConnectionKind::Outbound,
-                auth_type: input.auth_type,
-                allowed_hosts,
-                account_scope: input.account_scope,
-                origin: input.origin.unwrap_or(ConnectionOrigin::Vault),
-                external_source: input.external_source,
-                template_id: input.template_id,
-                protocol: input.protocol,
-                exposed_to: vec![],
-                trusted_for: vec![],
-                trusted_plugins: vec![],
-                legacy_ai_visible: None,
-                legacy_ai_trusted: None,
-                slots: Vec::new(),
-                last_used_at: None,
-                last_secret_updated_at: None,
-                display_name: None,
-                icon: None,
-                notes: input.notes,
-                created_at: now.clone(),
-                updated_at: now,
-            };
-            file.connections.push(connection.clone());
-            connection
-        }
-    };
-
-    connections_store::save(app, &file)?;
-    Ok(connection)
 }
 
 /// 全接続のメタデータ一覧を返す (secret は含まない)。
@@ -238,13 +63,10 @@ pub async fn vault_upsert_connection(
     input: ConnectionUpsert,
 ) -> VaultResult<Connection> {
     assert_main_window(&window)?;
-    upsert_metadata(&app, input)
+    service::upsert_metadata(&app, input)
 }
 
 /// 接続のメタデータと secret を 1 トランザクションで作成 / 更新する。
-///
-/// TOCTOU を避けるため、メタデータ保存 → keychain 書き込みを 1 コマンドにまとめる。
-/// keychain 書き込みに失敗したらメタデータ側の slot 登録もロールバックする。
 #[tauri::command]
 #[specta::specta]
 pub async fn vault_upsert_connection_with_secret(
@@ -255,39 +77,7 @@ pub async fn vault_upsert_connection_with_secret(
     secret: String,
 ) -> VaultResult<Connection> {
     assert_main_window(&window)?;
-    validate_slot(&slot)?;
-    if secret.len() < MIN_SECRET_LEN {
-        return Err(VaultError::SecretTooShort);
-    }
-
-    let connection = upsert_metadata(&app, input)?;
-    let conn_id = connection.id.clone();
-    let secret = SecretString::from(secret);
-
-    // keychain 書き込み。失敗したらメタデータは既に保存済みだが slot は未登録なので
-    // 「secret 未設定の接続」として残るだけで整合性は保たれる。
-    backend().store(&conn_id, &slot, &secret)?;
-
-    // slot をメタデータに登録する。
-    register_slot(&app, &conn_id, &slot)
-}
-
-/// slot をメタデータの `slots` 配列に登録し、更新後の接続を返す。
-fn register_slot(app: &tauri::AppHandle, conn_id: &str, slot: &str) -> VaultResult<Connection> {
-    let mut file = connections_store::load(app)?;
-    let connection = file
-        .connections
-        .iter_mut()
-        .find(|c| c.id == conn_id)
-        .ok_or(VaultError::ConnectionNotFound)?;
-    if !connection.slots.iter().any(|s| s == slot) {
-        connection.slots.push(slot.to_string());
-    }
-    connection.last_secret_updated_at = Some(now_millis());
-    connection.updated_at = now_millis();
-    let result = connection.clone();
-    connections_store::save(app, &file)?;
-    Ok(result)
+    service::upsert_with_secret(&app, input, &slot, secret)
 }
 
 /// 既存接続の secret を設定 / 入れ替える。
@@ -301,21 +91,7 @@ pub async fn vault_set_secret(
     secret: String,
 ) -> VaultResult<Connection> {
     assert_main_window(&window)?;
-    validate_connection_id(&id)?;
-    validate_slot(&slot)?;
-    if secret.len() < MIN_SECRET_LEN {
-        return Err(VaultError::SecretTooShort);
-    }
-
-    // 接続が存在することを確認する。
-    let file = connections_store::load(&app)?;
-    if !file.connections.iter().any(|c| c.id == id) {
-        return Err(VaultError::ConnectionNotFound);
-    }
-
-    let secret = SecretString::from(secret);
-    backend().store(&id, &slot, &secret)?;
-    register_slot(&app, &id, &slot)
+    service::set_secret(&app, &id, &slot, secret)
 }
 
 /// 接続の secret 設定状況を返す (値そのものは決して返さない)。
@@ -327,24 +103,7 @@ pub async fn vault_get_secret_status(
     id: String,
 ) -> VaultResult<SecretStatus> {
     assert_main_window(&window)?;
-    validate_connection_id(&id)?;
-    let file = connections_store::load(&app)?;
-    let connection = file
-        .connections
-        .iter()
-        .find(|c| c.id == id)
-        .ok_or(VaultError::ConnectionNotFound)?;
-
-    let backend = backend();
-    let mut slots = Vec::new();
-    for slot in &connection.slots {
-        let present = backend.load(&id, slot)?.is_some();
-        slots.push(SecretSlotStatus {
-            name: slot.clone(),
-            present,
-        });
-    }
-    Ok(SecretStatus { slots })
+    service::secret_status(&app, &id)
 }
 
 /// 接続の特定 slot の secret を削除する。
@@ -357,18 +116,7 @@ pub async fn vault_delete_secret(
     slot: String,
 ) -> VaultResult<()> {
     assert_main_window(&window)?;
-    validate_connection_id(&id)?;
-    validate_slot(&slot)?;
-
-    backend().delete(&id, &slot)?;
-
-    let mut file = connections_store::load(&app)?;
-    if let Some(connection) = file.connections.iter_mut().find(|c| c.id == id) {
-        connection.slots.retain(|s| s != &slot);
-        connection.updated_at = now_millis();
-        connections_store::save(&app, &file)?;
-    }
-    Ok(())
+    service::delete_secret(&app, &id, &slot)
 }
 
 /// 接続を削除する。全 slot の secret を keychain から消し、メタデータも削除する。
@@ -380,23 +128,7 @@ pub async fn vault_delete_connection(
     id: String,
 ) -> VaultResult<()> {
     assert_main_window(&window)?;
-    validate_connection_id(&id)?;
-
-    let mut file = connections_store::load(&app)?;
-    let Some(pos) = file.connections.iter().position(|c| c.id == id) else {
-        return Err(VaultError::ConnectionNotFound);
-    };
-
-    // secret を先に消す (途中 crash でも orphan メタデータより orphan secret の方が安全)。
-    let backend = backend();
-    let connection = &file.connections[pos];
-    for slot in &connection.slots {
-        backend.delete(&id, slot)?;
-    }
-
-    file.connections.remove(pos);
-    connections_store::save(&app, &file)?;
-    Ok(())
+    service::delete_connection(&app, &id)
 }
 
 /// 接続の開示先クラスを切り替える (#712 §6.1)。
@@ -406,33 +138,13 @@ pub async fn vault_set_exposed(
     app: tauri::AppHandle,
     window: tauri::Window,
     id: String,
-    principal_class: crate::vault::model::PrincipalClass,
+    principal_class: PrincipalClass,
     exposed: bool,
 ) -> VaultResult<()> {
     assert_main_window(&window)?;
-    validate_connection_id(&id)?;
-
-    let mut file = connections_store::load(&app)?;
-    let connection = file
-        .connections
-        .iter_mut()
-        .find(|c| c.id == id)
-        .ok_or(VaultError::ConnectionNotFound)?;
-    if exposed {
-        if !connection.exposed_to.contains(&principal_class) {
-            connection.exposed_to.push(principal_class);
-        }
-    } else {
-        connection.exposed_to.retain(|c| *c != principal_class);
-        // 開示を外したクラスの trust も意味を失うので同時に外す
-        connection.trusted_for.retain(|c| *c != principal_class);
-        if principal_class == crate::vault::model::PrincipalClass::Plugin {
-            connection.trusted_plugins.clear();
-        }
-    }
-    connection.updated_at = now_millis();
-    connections_store::save(&app, &file)?;
-    Ok(())
+    service::update_connection(&app, &id, |c| {
+        service::apply_exposed(c, principal_class, exposed)
+    })
 }
 
 /// 接続を「信頼済み」(確認なしで利用可) にするクラスを切り替える (#712 §6.2)。
@@ -444,35 +156,19 @@ pub async fn vault_set_trusted(
     app: tauri::AppHandle,
     window: tauri::Window,
     id: String,
-    principal_class: crate::vault::model::PrincipalClass,
+    principal_class: PrincipalClass,
     trusted: bool,
 ) -> VaultResult<()> {
     assert_main_window(&window)?;
-    validate_connection_id(&id)?;
-
-    let mut file = connections_store::load(&app)?;
-    let connection = file
-        .connections
-        .iter_mut()
-        .find(|c| c.id == id)
-        .ok_or(VaultError::ConnectionNotFound)?;
-    if trusted {
-        if !connection.trusted_for.contains(&principal_class) {
-            connection.trusted_for.push(principal_class);
-        }
-    } else {
-        connection.trusted_for.retain(|c| *c != principal_class);
-    }
-    connection.updated_at = now_millis();
-    connections_store::save(&app, &file)?;
-    Ok(())
+    service::update_connection(&app, &id, |c| {
+        service::apply_trusted(c, principal_class, trusted)
+    })
 }
 
 /// 接続を「信頼済み」にするプラグイン個体を切り替える。
 ///
 /// plugin クラスの trust はクラス一括 (`trusted_for`) にせず個体単位で持つ —
 /// 1 つのウィジェットの確認同意が全プラグイン / Play / Page に波及しない。
-/// `name` は帰属表示用スナップショット (再信頼で最新の名前に更新される)。
 #[tauri::command]
 #[specta::specta]
 pub async fn vault_set_trusted_plugin(
@@ -484,26 +180,9 @@ pub async fn vault_set_trusted_plugin(
     trusted: bool,
 ) -> VaultResult<()> {
     assert_main_window(&window)?;
-    validate_connection_id(&id)?;
-
-    let mut file = connections_store::load(&app)?;
-    let connection = file
-        .connections
-        .iter_mut()
-        .find(|c| c.id == id)
-        .ok_or(VaultError::ConnectionNotFound)?;
-    connection.trusted_plugins.retain(|p| p.id != plugin_id);
-    if trusted {
-        connection
-            .trusted_plugins
-            .push(crate::vault::model::TrustedPlugin {
-                id: plugin_id,
-                name,
-            });
-    }
-    connection.updated_at = now_millis();
-    connections_store::save(&app, &file)?;
-    Ok(())
+    service::update_connection(&app, &id, |c| {
+        service::apply_trusted_plugin(c, plugin_id, name, trusted)
+    })
 }
 
 /// 登録済み接続を使って HTTP リクエストを実行する。
@@ -523,20 +202,8 @@ pub async fn vault_fetch(
 ) -> VaultResult<VaultFetchResponse> {
     assert_main_window(&window)?;
     let response = fetch::vault_fetch(&app, &id, request).await?;
-    touch_last_used(&app, &id);
+    service::touch_last_used(&app, &id);
     Ok(response)
-}
-
-/// 接続の疎通テスト結果。
-#[derive(Debug, Serialize, specta::Type)]
-#[serde(rename_all = "camelCase")]
-pub struct VaultTestResult {
-    /// HTTP ステータス (リクエストが届いた場合)。
-    pub status: Option<u16>,
-    /// 2xx / 3xx なら true。
-    pub ok: bool,
-    /// 失敗時の理由 (SSRF / timeout / DNS など)。secret は含まない。
-    pub error: Option<String>,
 }
 
 /// 接続の疎通テスト。baseUrl への GET (または指定パス) を 1 回実行する。
@@ -549,39 +216,10 @@ pub async fn vault_test_connection(
     test_path: Option<String>,
 ) -> VaultResult<VaultTestResult> {
     assert_main_window(&window)?;
-    validate_connection_id(&id)?;
-
-    let request = VaultFetchRequest {
-        path: test_path.unwrap_or_else(|| "/".to_string()),
-        method: Some("GET".to_string()),
-        headers: None,
-        body: None,
-        timeout_ms: Some(10_000),
-        slot: None,
-    };
-
-    match fetch::vault_fetch(&app, &id, request).await {
-        Ok(resp) => Ok(VaultTestResult {
-            status: Some(resp.status),
-            ok: (200..400).contains(&resp.status),
-            error: None,
-        }),
-        Err(e) => Ok(VaultTestResult {
-            status: None,
-            ok: false,
-            error: Some(e.to_string()),
-        }),
-    }
+    service::test_connection(&app, &id, test_path).await
 }
 
 /// AI プロバイダーの API キーを Vault 接続へ移行する (#564 後続)。
-///
-/// 旧来 `ai.<provider>` キーチェーンに保存していた AI API キーを、Vault の
-/// 接続 (`origin = External`, `externalSource = "ai-provider"`) に移し替える。
-/// 移行後、旧キーチェーンエントリーは削除する。
-///
-/// キーチェーンに該当 provider のエントリーが無い場合は `None` を返す
-/// (移行対象なし)。フロント側は返り値の接続 id を `ai.json5` に記録する。
 #[tauri::command]
 #[specta::specta]
 pub async fn ai_migrate_provider_to_vault(
@@ -593,58 +231,5 @@ pub async fn ai_migrate_provider_to_vault(
     protocol: ConnectionProtocol,
 ) -> VaultResult<Option<Connection>> {
     assert_main_window(&window)?;
-
-    let api_key = crate::commands::ai::read_ai_api_key(&provider)
-        .map_err(|e| VaultError::InvalidInput {
-            message: e.to_string(),
-        })?;
-    let Some(api_key) = api_key.filter(|k| !k.is_empty()) else {
-        return Ok(None);
-    };
-
-    // AI チャットの認証は `ai_chat.rs` 側で protocol 別に注入するが、`vault_fetch`
-    // 経由でも使えるよう authType も protocol に合わせて設定しておく。
-    let auth_type = match protocol {
-        ConnectionProtocol::Anthropic => AuthType::Header {
-            name: "x-api-key".to_string(),
-        },
-        ConnectionProtocol::OpenaiCompat => AuthType::Bearer,
-    };
-
-    let connection = upsert_metadata(
-        &app,
-        ConnectionUpsert {
-            id: None,
-            name,
-            base_url,
-            auth_type,
-            allowed_hosts: Vec::new(),
-            account_scope: None,
-            notes: None,
-            template_id: None,
-            protocol: Some(protocol),
-            origin: Some(ConnectionOrigin::External),
-            external_source: Some("ai-provider".to_string()),
-        },
-    )?;
-
-    let conn_id = connection.id.clone();
-    backend().store(&conn_id, "primary", &SecretString::from(api_key))?;
-    let connection = register_slot(&app, &conn_id, "primary")?;
-
-    // 旧キーチェーンエントリーを削除する。失敗しても移行自体は成功扱い。
-    let _ = notecli::keychain::delete_token(&crate::commands::ai::ai_keychain_id(&provider));
-
-    Ok(Some(connection))
-}
-
-/// 接続の `last_used_at` を現在時刻で更新する (ベストエフォート、失敗は無視)。
-fn touch_last_used(app: &tauri::AppHandle, id: &str) {
-    let Ok(mut file) = connections_store::load(app) else {
-        return;
-    };
-    if let Some(connection) = file.connections.iter_mut().find(|c| c.id == id) {
-        connection.last_used_at = Some(now_millis());
-        let _ = connections_store::save(app, &file);
-    }
+    service::migrate_ai_provider(&app, &provider, name, base_url, protocol)
 }
